Handle subshells correctly when entering a new environment at the same time
Previously we left and entered a new state before cloning the
environment.  This caused issues when we not only created a subshell,
but also changed a directory that changed scope of an env config.  What
we ended up doing is reusing parent shell's state, which caused a full
clean load in parent shell when we exited the child shell.  As a
consequence the environment actions we replayed on top of an environment
where they were already applied.

This new implementation correctly handles this case by cloning the
environment when we enter a new subshell, unless we are entering a
different environment.  All we have to do now is make sure that we left
parent state file where it was.  If we have undiscriminately removed old
states we would force parent shell to enter a state without leaving it
properly.  In such case we would end up being in the same position as
before.