Installed `fzy` and `ripgrep` on Debian
We need to check that these are available on Buntish systems as well
before moving them to list of shared packages.

For `fzy` there is a command smart constructor provided for
Command Wrapper's `cd` subcommand.  It's algorithm seems to work better
than `fzf`'s, however, it's somewhat limited regarding features.