Supporting Neovim remote functionality
When enabled `yx jmp` can open a file in remote Neovim by calling it's
API.  This functionality can be controlled by `--[no-]neovim-remote`
options.

Another enhancement is that we can now detect that `yx jmp` is running
in a Vim/Neovim terminal.  In such case we are unable to access
scrollback buffer at the moment.